Downloaded 5 photos for the featuredshows.component.
Created a featured_shows_model.ts to be able to collect data from the backend.

Created a featuredshows on firebase and added a featuredshows.service.ts to be able to collect all the data from the backend.

Currently finding how to apply the data from the backend onto the featuredshows.component.ts